+++ test/3_fund_owner.js
@@ -24,21 +24,21 @@ const DummyRariFundManager = artifacts.require("DummyRariFundManager");
 // These tests expect the owner and the fund rebalancer of RariFundController and RariFundManager to be set to process.env.DEVELOPMENT_ADDRESS
 contract("RariFundController, RariFundManager", accounts => {
   it("should upgrade the fund manager owner", async () => {
-    let fundManagerInstance = await RariFundManager.deployed();
+    let fundManagerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundManager.at(process.env.UPGRADE_FUND_MANAGER_ADDRESS) : RariFundManager.deployed());
 
     // RariFundManager.transferOwnership()
     await fundManagerInstance.transferOwnership(process.env.DEVELOPMENT_ADDRESS_SECONDARY, { from: process.env.DEVELOPMENT_ADDRESS });
 
     // Test disabling and enabling the fund from the new owner address
-    await fundManagerInstance.disableFund({ from: process.env.DEVELOPMENT_ADDRESS_SECONDARY });
-    await fundManagerInstance.enableFund({ from: process.env.DEVELOPMENT_ADDRESS_SECONDARY });
+    await fundManagerInstance.setFundDisabled(true, { from: process.env.DEVELOPMENT_ADDRESS_SECONDARY });
+    await fundManagerInstance.setFundDisabled(false, { from: process.env.DEVELOPMENT_ADDRESS_SECONDARY });
 
     // Transfer ownership back
     await fundManagerInstance.transferOwnership(process.env.DEVELOPMENT_ADDRESS, { from: process.env.DEVELOPMENT_ADDRESS_SECONDARY });
   });
 
   it("should upgrade the fund controller owner", async () => {
-    let fundControllerInstance = await RariFundController.deployed();
+    let fundControllerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundController.at(process.env.UPGRADE_FUND_CONTROLLER_ADDRESS) : RariFundController.deployed());
 
     // RariFundManager.transferOwnership()
     await fundControllerInstance.transferOwnership(process.env.DEVELOPMENT_ADDRESS_SECONDARY, { from: process.env.DEVELOPMENT_ADDRESS });
@@ -52,14 +52,15 @@ contract("RariFundController, RariFundManager", accounts => {
   });
 
   it("should disable and re-enable the fund", async () => {
-    let fundControllerInstance = await RariFundController.deployed();
-    let fundManagerInstance = await RariFundManager.deployed();
-    let fundTokenInstance = await RariFundToken.deployed();
-    let fundPriceConsumerInstance = await RariFundPriceConsumer.deployed();
+    let fundControllerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundController.at(process.env.UPGRADE_FUND_CONTROLLER_ADDRESS) : RariFundController.deployed());
+    let fundManagerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundManager.at(process.env.UPGRADE_FUND_MANAGER_ADDRESS) : RariFundManager.deployed());
+    if (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0) RariFundManager.address = process.env.UPGRADE_FUND_MANAGER_ADDRESS;
+    let fundTokenInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundToken.at(process.env.UPGRADE_FUND_TOKEN_ADDRESS) : RariFundToken.deployed());
+    let fundPriceConsumerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundPriceConsumer.at(process.env.UPGRADE_FUND_PRICE_CONSUMER_ADDRESS) : RariFundPriceConsumer.deployed());
 
     // Disable the fund (via RariFundController and RariFundManager)
     await fundControllerInstance.disableFund({ from: process.env.DEVELOPMENT_ADDRESS });
-    await fundManagerInstance.disableFund({ from: process.env.DEVELOPMENT_ADDRESS });
+    await fundManagerInstance.setFundDisabled(true, { from: process.env.DEVELOPMENT_ADDRESS });
 
     // TODO: Check _fundDisabled (no way to do this as of now)
     
@@ -98,7 +99,7 @@ contract("RariFundController, RariFundManager", accounts => {
     }
 
     // Re-enable the fund (via RariFundManager and RariFundController)
-    await fundManagerInstance.enableFund({ from: process.env.DEVELOPMENT_ADDRESS, nonce: await web3.eth.getTransactionCount(process.env.DEVELOPMENT_ADDRESS) });
+    await fundManagerInstance.setFundDisabled(false, { from: process.env.DEVELOPMENT_ADDRESS, nonce: await web3.eth.getTransactionCount(process.env.DEVELOPMENT_ADDRESS) });
     await fundControllerInstance.enableFund({ from: process.env.DEVELOPMENT_ADDRESS, nonce: await web3.eth.getTransactionCount(process.env.DEVELOPMENT_ADDRESS) });
 
     // TODO: Check _fundDisabled (no way to do this as of now)
@@ -119,8 +120,8 @@ contract("RariFundController, RariFundManager", accounts => {
   });
 
   it("should upgrade the fund rebalancer", async () => {
-    let fundControllerInstance = await RariFundController.deployed();
-    let fundManagerInstance = await RariFundManager.deployed();
+    let fundControllerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundController.at(process.env.UPGRADE_FUND_CONTROLLER_ADDRESS) : RariFundController.deployed());
+    let fundManagerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundManager.at(process.env.UPGRADE_FUND_MANAGER_ADDRESS) : RariFundManager.deployed());
 
     // Set fund rebalancer addresses
     await fundControllerInstance.setFundRebalancer(process.env.DEVELOPMENT_ADDRESS_SECONDARY, { from: process.env.DEVELOPMENT_ADDRESS });
@@ -142,9 +143,10 @@ contract("RariFundController, RariFundManager", accounts => {
 
 contract("RariFundManager", accounts => {
   it("should upgrade the FundManager implementation to a copy of its code", async () => {
-    let fundControllerInstance = await RariFundController.deployed();
-    let fundManagerInstance = await RariFundManager.deployed();
-    
+    let fundControllerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundController.at(process.env.UPGRADE_FUND_CONTROLLER_ADDRESS) : RariFundController.deployed());
+    let fundManagerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundManager.at(process.env.UPGRADE_FUND_MANAGER_ADDRESS) : RariFundManager.deployed());
+    if (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0) RariFundManager.address = process.env.UPGRADE_FUND_MANAGER_ADDRESS;
+
     // Approve and deposit tokens to the fund (using DAI as an example)
     var amountBN = web3.utils.toBN(10 ** (currencies["DAI"].decimals - 1));
     var erc20Contract = new web3.eth.Contract(erc20Abi, currencies["DAI"].tokenAddress);
@@ -162,12 +164,14 @@ contract("RariFundManager", accounts => {
 
     // Disable FundController and original FundManager
     await fundControllerInstance.disableFund({ from: process.env.DEVELOPMENT_ADDRESS });
-    await fundManagerInstance.disableFund({ from: process.env.DEVELOPMENT_ADDRESS });
+    await fundManagerInstance.setFundDisabled(true, { from: process.env.DEVELOPMENT_ADDRESS });
 
     // TODO: Check _fundDisabled (no way to do this as of now)
 
     // Upgrade FundManager
+    if (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0) RariFundManager.class_defaults.from = process.env.UPGRADE_FUND_OWNER_ADDRESS;
     var newFundManagerInstance = await upgradeProxy(RariFundManager.address, RariFundManager, { unsafeAllowCustomTypes: true });
+    if (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0) RariFundManager.class_defaults.from = process.env.DEVELOPMENT_ADDRESS;
 
     // Check balance of new FundManager
     let newRawFundBalance = await newFundManagerInstance.getRawFundBalance.call();
@@ -181,9 +185,10 @@ contract("RariFundManager", accounts => {
 
 contract("RariFundManager", accounts => {
   it("should upgrade the proxy and implementation of FundManager to new code", async () => {
-    let fundControllerInstance = await RariFundController.deployed();
-    let fundManagerInstance = await RariFundManager.deployed();
-    
+    let fundControllerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundController.at(process.env.UPGRADE_FUND_CONTROLLER_ADDRESS) : RariFundController.deployed());
+    let fundManagerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundManager.at(process.env.UPGRADE_FUND_MANAGER_ADDRESS) : RariFundManager.deployed());
+    if (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0) RariFundManager.address = process.env.UPGRADE_FUND_MANAGER_ADDRESS;
+
     // Approve and deposit tokens to the fund (using DAI as an example)
     var amountBN = web3.utils.toBN(10 ** (currencies["DAI"].decimals - 1));
     var erc20Contract = new web3.eth.Contract(erc20Abi, currencies["DAI"].tokenAddress);
@@ -196,7 +201,7 @@ contract("RariFundManager", accounts => {
 
     // Disable FundController and original FundManager
     await fundControllerInstance.disableFund({ from: process.env.DEVELOPMENT_ADDRESS });
-    await fundManagerInstance.disableFund({ from: process.env.DEVELOPMENT_ADDRESS });
+    await fundManagerInstance.setFundDisabled(true, { from: process.env.DEVELOPMENT_ADDRESS });
 
     // TODO: Check _fundDisabled (no way to do this as of now)
 
@@ -212,9 +217,10 @@ contract("RariFundManager", accounts => {
 
 contract("RariFundController", accounts => {
   it("should upgrade the FundController to a copy of its code", async () => {
-    let fundControllerInstance = await RariFundController.deployed();
-    let fundManagerInstance = await RariFundManager.deployed();
-    
+    let fundControllerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundController.at(process.env.UPGRADE_FUND_CONTROLLER_ADDRESS) : RariFundController.deployed());
+    let fundManagerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundManager.at(process.env.UPGRADE_FUND_MANAGER_ADDRESS) : RariFundManager.deployed());
+    if (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0) RariFundManager.address = process.env.UPGRADE_FUND_MANAGER_ADDRESS;
+
     // Approve and deposit tokens to the fund (using DAI as an example)
     var amountBN = web3.utils.toBN(10 ** (currencies["DAI"].decimals - 1));
     var erc20Contract = new web3.eth.Contract(erc20Abi, currencies["DAI"].tokenAddress);
@@ -232,7 +238,7 @@ contract("RariFundController", accounts => {
 
     // Disable FundController and original FundManager
     await fundControllerInstance.disableFund({ from: process.env.DEVELOPMENT_ADDRESS });
-    await fundManagerInstance.disableFund({ from: process.env.DEVELOPMENT_ADDRESS });
+    await fundManagerInstance.setFundDisabled(true, { from: process.env.DEVELOPMENT_ADDRESS });
 
     // TODO: Check _fundDisabled (no way to do this as of now)
 
@@ -256,9 +262,10 @@ contract("RariFundController", accounts => {
 
 contract("RariFundController", accounts => {
   it("should upgrade the FundController to new code", async () => {
-    let fundControllerInstance = await RariFundController.deployed();
-    let fundManagerInstance = await RariFundManager.deployed();
-    
+    let fundControllerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundController.at(process.env.UPGRADE_FUND_CONTROLLER_ADDRESS) : RariFundController.deployed());
+    let fundManagerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundManager.at(process.env.UPGRADE_FUND_MANAGER_ADDRESS) : RariFundManager.deployed());
+    if (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0) RariFundManager.address = process.env.UPGRADE_FUND_MANAGER_ADDRESS;
+
     // Approve and deposit tokens to the fund (using DAI as an example)
     var amountBN = web3.utils.toBN(10 ** (currencies["DAI"].decimals - 1));
     var erc20Contract = new web3.eth.Contract(erc20Abi, currencies["DAI"].tokenAddress);
@@ -276,7 +283,7 @@ contract("RariFundController", accounts => {
 
     // Disable FundController and original FundManager
     await fundControllerInstance.disableFund({ from: process.env.DEVELOPMENT_ADDRESS });
-    await fundManagerInstance.disableFund({ from: process.env.DEVELOPMENT_ADDRESS });
+    await fundManagerInstance.setFundDisabled(true, { from: process.env.DEVELOPMENT_ADDRESS });
 
     // TODO: Check _fundDisabled (no way to do this as of now)
 
@@ -299,10 +306,12 @@ contract("RariFundController", accounts => {
 
 contract("RariFundToken", accounts => {
   it("should upgrade the FundToken to a copy of its code", async () => {
-    let fundControllerInstance = await RariFundController.deployed();
-    let fundManagerInstance = await RariFundManager.deployed();
-    let fundTokenInstance = await RariFundToken.deployed();
-    
+    let fundControllerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundController.at(process.env.UPGRADE_FUND_CONTROLLER_ADDRESS) : RariFundController.deployed());
+    let fundManagerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundManager.at(process.env.UPGRADE_FUND_MANAGER_ADDRESS) : RariFundManager.deployed());
+    if (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0) RariFundManager.address = process.env.UPGRADE_FUND_MANAGER_ADDRESS;
+    let fundTokenInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundToken.at(process.env.UPGRADE_FUND_TOKEN_ADDRESS) : RariFundToken.deployed());
+    if (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0) RariFundToken.address = process.env.UPGRADE_FUND_TOKEN_ADDRESS;
+
     // Approve and deposit tokens to the fund (using DAI as an example)
     var amountBN = web3.utils.toBN(10 ** (currencies["DAI"].decimals - 1));
     var erc20Contract = new web3.eth.Contract(erc20Abi, currencies["DAI"].tokenAddress);
@@ -317,12 +326,14 @@ contract("RariFundToken", accounts => {
 
     // Disable FundController and original FundManager
     await fundControllerInstance.disableFund({ from: process.env.DEVELOPMENT_ADDRESS });
-    await fundManagerInstance.disableFund({ from: process.env.DEVELOPMENT_ADDRESS });
+    await fundManagerInstance.setFundDisabled(true, { from: process.env.DEVELOPMENT_ADDRESS });
 
     // TODO: Check _fundDisabled (no way to do this as of now)
 
     // Upgrade FundToken
+    if (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0) RariFundToken.class_defaults.from = process.env.UPGRADE_FUND_OWNER_ADDRESS;
     var newFundTokenInstance = await upgradeProxy(RariFundToken.address, RariFundToken);
+    if (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0) RariFundToken.class_defaults.from = process.env.DEVELOPMENT_ADDRESS;
 
     // Check balances
     let newRawFundBalance = await fundManagerInstance.getRawFundBalance.call();