+++ truffle-config.js
@@ -15,7 +15,7 @@ module.exports = {
     live: {
       provider: function() {
         var keys = [process.env.LIVE_DEPLOYER_PRIVATE_KEY];
-        if (process.env.UPGRADE_FUND_OWNER_PRIVATE_KEY) keys.push(process.env.UPGRADE_FUND_OWNER_PRIVATE_KEY);
+        if (process.env.LIVE_UPGRADE_FUND_OWNER_PRIVATE_KEY) keys.push(process.env.LIVE_UPGRADE_FUND_OWNER_PRIVATE_KEY);
         return new HDWalletProvider(keys, process.env.LIVE_WEB3_PROVIDER_URL);
       },
       network_id: 1,